Loosen up on GrumPhp
The version range on GrumPhp is too strict, forcing it to remain on an old version of more than a year ago. This is due to the difference in handling packages that are still in version 0: using `^0.11` will prevent it from going up minor versions, while `~0.11` will allow it. Only packages that have a version 1 or higher are allowed to advance minor versions when using `^`.