OPENEUROPA-1535: Tighten up the coding standards for libraries.
The PSR-2 standard does not explicitly cover how comments, multi-line
statements and array declarations that span multiple lines should be
formatted. We require these to be indented with 4 spaces, so that they
are consistent with the rest of the code base.

The PSR-2 standard doesn't cover how concatenation operators are to be
formatted, causing inconsistent formatting to be ignored. In order to
rectify this in a future proof way we already borrow the rules from the
upcoming PSR-12 standard.