Do not force class elements to be ordered alphabetically and per visibility
The `php-cs-fixer` rule that forces class elements to be ordered alphabetically and per visibility is preventing developers from grouping related elements. For example forcing constants to be alphabetical can cause a group of related constants to have an unrelated constant forced in between them. The visibility ordering is equally counterproductive: for example a public method that defers its logic to a protected recursive method is now forced far apart at opposite sides of the class.